+++ build_modules/lib/src/modules.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
+import 'package:graphs/graphs.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import 'errors.dart';
@@ -113,8 +114,8 @@ class Module extends Object with _$ModuleSerializerMixin {
   /// Generated factory constructor.
   factory Module.fromJson(Map<String, dynamic> json) => _$ModuleFromJson(json);
 
-  /// Computes the [primarySource]s of all [Module]s that are transitively
-  /// depended on by this module.
+  /// Returns all [Module]s in the transitive dependencies of this module in
+  /// reverse dependency order.
   ///
   /// Throws a [MissingModulesException] if there are any missing modules. This
   /// typically means that somebody is trying to import a non-existing file.
@@ -141,7 +142,11 @@ class Module extends Object with _$ModuleSerializerMixin {
       throw await MissingModulesException.create(this, missingModuleSources,
           transitiveDeps.values.toList()..add(this), reader);
     }
-    return transitiveDeps.values.toList();
+    var orderedModules = stronglyConnectedComponents<AssetId, Module>(
+        transitiveDeps.values,
+        (m) => m.primarySource,
+        (m) => m.directDependencies.map((s) => transitiveDeps[s]));
+    return orderedModules.map((c) => c.single).toList();
   }
 }