+++ build_modules/test/fixtures/b/lib/b_second_import_to_a_no_cycle.dart
@@ -1,0 +1,2 @@
+// ignore_for_file: unused_import
+import 'package:a/a_no_cycle.dart';